+++ OOP/src/oop/model/Transaction.java
@@ -1,6 +1,9 @@
 package oop.model;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 
 
@@ -10,10 +13,9 @@ import lombok.AllArgsConstructor;
 public class Transaction {
     private String cardNumber;
     private String type;         // "DEPOSIT" or "WITHDRAW"
-    private double amount;
+    private BigDecimal amount;
     private String date;
 
-    // constructor, getters, setters, toFileString()
     
     public String toFileString() {
         return cardNumber + "|" + type + "|" + amount + "|" + date;
